+++ src/VBox/Devices/Graphics/DevVGA-SVGA3d-info.cpp
@@ -4,7 +4,7 @@
  */
 
 /*
- * Copyright (C) 2013-2024 Oracle and/or its affiliates.
+ * Copyright (C) 2013-2025 Oracle and/or its affiliates.
  *
  * This file is part of VirtualBox base platform packages, as
  * available from https://www.virtualbox.org.
@@ -1477,9 +1477,9 @@ static void vmsvga3dInfoContextWorkerOne(PCDBGFINFOHLP pHlp, PVMSVGA3DCONTEXT pC
                             FLOAT_FMT_ARGS(pContext->state.aLightData[i].data.phi));
         }
 
-    for (uint32_t i = 0; i < RT_ELEMENTS(pContext->state.aRenderTargets); i++)
+    for (size_t i = 0; i < RT_ELEMENTS(pContext->state.aRenderTargets); i++)
         if (pContext->state.aRenderTargets[i] != SVGA3D_INVALID_ID)
-            pHlp->pfnPrintf(pHlp, "aRenderTargets[%s/%u] = %#x (%d)\n",
+            pHlp->pfnPrintf(pHlp, "aRenderTargets[%s/%zu] = %#x (%d)\n",
                             i < RT_ELEMENTS(g_apszRenderTargets) ? g_apszRenderTargets[i] : "UNKNOWN", i,
                             pContext->state.aRenderTargets[i], pContext->state.aRenderTargets[i]);
