+++ src/VBox/Frontends/VBoxManage/VBoxManageGuestProp.cpp
@@ -4,7 +4,7 @@
  */
 
 /*
- * Copyright (C) 2006-2024 Oracle and/or its affiliates.
+ * Copyright (C) 2006-2025 Oracle and/or its affiliates.
  *
  * This file is part of VirtualBox base platform packages, as
  * available from https://www.virtualbox.org.
@@ -511,7 +511,7 @@ static RTEXITCODE handleWaitGuestProperty(HandlerArg *a)
         }
 
         ComPtr<IEvent> ev;
-        hrc = es->GetEvent(listener, cMsWait, ev.asOutParam());
+        hrc = es->GetEvent(listener, (LONG)cMsWait, ev.asOutParam());
         if (ev) /** @todo r=andy Why not using SUCCEEDED(hrc) here? */
         {
             VBoxEventType_T aType;
